pyproject.toml
[tool.poetry]
name = "bluebird"
version = "0.1.0"
description = ""
authors = ["Masahiro Ishii <marufeuille@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1,<4.0"
bluepy = "^1.3.0"
pandas = "^1.2.4"
matplotlib = "^3.4.2"
google-cloud-storage = "^1.38.0"
slackweb = "^1.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
